AIIncidentIntelligencePlatform/llm-layer/requirements.txt
fastapi>=0.100.0
uvicorn>=0.22.0
kafka-python>=2.0.2
openai>=1.54.0
python-dotenv>=1.0.0
prometheus-client>=0.19.0
